refractored updates based off review
